Changes to append .R into post processing filter name in GWS instead of GrapleR. List of Post processing filters returned without extention

+++ gws.py
@@ -18,15 +18,17 @@ import tarfile
 from pymongo import MongoClient
 from os import listdir
 import json
+import itertools
 
 # global variables and paths
 base_working_path = '/home/grapleadmin/grapleService/'
-
+ 
 base_upload_path = base_working_path + 'static'
 base_graple_path = base_working_path + 'GRAPLE_SCRIPTS'
 
 app = Flask(__name__, static_url_path='', static_folder = base_upload_path)
 app.config['CELERY_BROKER_URL'] = 'amqp://'
+app.config['MAX_CONTENT_LENGTH'] = 4*1024*1024*1024
 #app.config['CELERY_RESULT_BACKEND'] = 'ampq'
 celery = Celery(app.name, broker=app.config['CELERY_BROKER_URL'])
 celery.conf.update(app.config)
@@ -66,8 +68,9 @@ def setup_graple(path,filename, rscript):
     if(rscript):
         filename = rscript 
         os.chdir("Scripts")
-        shutil.copy(os.path.join(topdir, "Filters", filename),os.getcwd())
-        os.rename(filename, 'PostProcessFilter.R')
+        if os.path.isfile(os.path.join(topdir, "Filters", filename)):
+            shutil.copy(os.path.join(topdir, "Filters", filename),os.getcwd())
+            os.rename(filename, 'PostProcessFilter.R')
         filterParamsDir = os.path.join(topdir, "Sims", "FilterParams")
         if(os.path.exists(filterParamsDir)):
             shutil.copy(os.path.join(filterParamsDir, "FilterParams.json"), os.getcwd())
@@ -153,7 +156,9 @@ def handle_linearsweep_run(task, rscript):
     if (task.split('$')[0] == "run_linear_sweep"):
         dir_name = task.split('$')[1]
         filename = task.split('$')[2]
-        current_dir = os.path.join(os.getcwd(), dir_name)
+        print task
+        current_dir = dir_name
+        print current_dir
         base_folder = os.path.join(dir_name,'base_folder')
         # unzip the zip file and read job_desc.csv
         os.chdir(base_folder)
@@ -196,12 +201,19 @@ def handle_linearsweep_run(task, rscript):
             Sims_dir=os.path.join(dir_name,'Sims')
             shutil.rmtree(Sims_dir)
             os.mkdir(Sims_dir)
-        print base_iterations+1
-        for j in range(1,base_iterations+1):
+        varcomb = []
+        numcomb = []
+        sim_no = 1
+        for jsonfilename, variables in columns.iteritems():
+            for variable, stepValues in variables.iteritems():
+                varcomb.append(jsonfilename +',' + variable + ',' + stepValues[2] + ',' + stepValues[1])
+                numcomb.append(stepValues[0])
+
+        for comb in itertools.product(*numcomb):
             os.chdir(Sims_dir)
-            new_dir="Sim"+ str(j)
+            new_dir="Sim"+ str(sim_no)
             if not os.path.exists(new_dir):
-                summary.append(["sim_"+str(j)])
+                summary.append(["sim_"+str(sim_no)])
                 os.mkdir(new_dir)
                 os.chdir(new_dir)
                 shutil.copy(os.path.join(dir_name,'base_folder',filename),os.getcwd())
@@ -210,36 +222,34 @@ def handle_linearsweep_run(task, rscript):
                 os.remove("job_desc.json")
             else:
                 os.chdir(new_dir)
-            for key in columns.keys():
-                base_file = key             
+            for i in range(len(varcomb)):
+                var_list = varcomb[i].split(",")
+                base_file = var_list[0]
+                field = var_list[1]
+                operation = var_list[3]
+                delta = comb[i]
                 data = pandas.read_csv(base_file)
                 data = data.rename(columns=lambda x: x.strip())
-                for field in columns[base_file].keys():
-                    if (((" "+field) in data.columns) or (field in data.columns)):
-                    # handle variations in filed names in csv file, some field names have leading spaces.
-                        if " "+field in data.columns:
-                            field_modified = " "+field
-                        else:
-                            field_modified = field
-                        stepValue = j % int(columns[base_file][field][3])
-                        if stepValue == 0:
-                            stepValue = columns[base_file][field][3]
-                        delta = columns[base_file][field][0][stepValue-1]
-                        if (columns[base_file][field][1]=="add"):
-                            data[field_modified]=data[field_modified].apply(lambda val:val+delta)
-                        elif (columns[base_file][field][1]=="sub"):
-                            data[base_file][field_modified]=data[field_modified].apply(lambda val:val-delta)
-                        elif (columns[base_file][field][1]=="mul"):
-                            data[field_modified]=data[field_modified].apply(lambda val:val*delta)
-                        elif (columns[base_file][field][1]=="div"):
-                            data[field_modified]=data[field_modified].apply(lambda val:val/delta)
-                        # make note of modified changes in a list datastructure
-                        summary[j-1].append(field) # append column_name
-                        summary[j-1].append(columns[base_file][field][2]) # append distribution
-                        summary[j-1].append(columns[base_file][field][1]) # append operation
-                        summary[j-1].append(str(delta)) # append delta
-                    # at this point the dataframe has been modified, write back to csv.
+                if (((" "+field) in data.columns) or (field in data.columns)):
+                # handle variations in filed names in csv file, some field names have leading spaces.
+                    if " "+field in data.columns:
+                        field_modified = " "+field
+                    else:
+                        field_modified = field
+                if (operation=="add"):
+                    data[field_modified]=data[field_modified].apply(lambda val:val+delta)
+                elif (operation=="sub"):
+                    data[base_file][field_modified]=data[field_modified].apply(lambda val:val-delta)
+                elif (operation=="mul"):
+                    data[field_modified]=data[field_modified].apply(lambda val:val*delta)
+                elif (operation=="div"):
+                    data[field_modified]=data[field_modified].apply(lambda val:val/delta)
+                summary[sim_no - 1].append(field)
+                summary[sim_no - 1].append(columns[base_file][field][2]) 
+                summary[sim_no - 1].append(columns[base_file][field][1]) 
+                summary[sim_no - 1].append(str(delta))
                 data.to_csv(base_file,index=False)
+            sim_no += 1
         print str(summary)
         # write summary of modifications to a file.
         result_summary = open(os.path.join(base_folder,"sim_summary.csv"),'wb')
@@ -319,9 +329,9 @@ def handle_special_job(task, rscript):
                 os.remove(filename)
                 os.remove("job_desc.json")
             else:                
-                os.chdir(new_dir)
+                os.chdir(new_dir)    
             for key in columns.keys():
-                base_file = key
+                base_file = key             
                 data = pandas.read_csv(base_file)
                 data = data.rename(columns=lambda x: x.strip())  
                 for field in columns[base_file].keys():
@@ -392,6 +402,7 @@ def special_batch(filtername):
         # have to submit job to celery here--below method has to be handled by celery worker
         task_desc = "graple_special_batch"+"$"+dir_name+"$"+filename
         if(filtername):
+            filtername += '.R'
             doTask.delay(task_desc, filtername)
         else:
             doTask.delay(task_desc)
@@ -419,6 +430,7 @@ def linear_sweep(filtername):
         # have to submit job to celery here--below method has to be handled by celery worker
         task_desc = "run_linear_sweep"+"$"+dir_name+"$"+filename
         if(filtername):
+            filtername += '.R'
             doTask.delay(task_desc, filtername)
         else:
             doTask.delay(task_desc)
@@ -445,27 +457,6 @@ def abort_job(uid):
         status["curr_status"]=Abort_Job(dir_name)
         return jsonify(status)
 
-@app.route('/TriggerSimulation/<sweepstring>', methods=['GET','POST'])
-def return_file(sweepstring):
-    global base_upload_path
-    base_params = sweepstring.split("*")
-    uid=base_params[0]
-    if request.method == 'GET':
-        dir_name = os.path.join(base_upload_path,str(uid))
-        os.chdir(dir_name)
-        contents = {}
-        for each in os.listdir(dir_name):
-            contents[each] = os.path.getsize(each)
-        return jsonify(contents)
-    
-    if request.method == 'POST':
-        dir_name = os.path.join(base_upload_path,str(uid))
-        # put the task in the queue in the required format
-        task_desc = "run_sweep"+"$"+dir_name+"$"+sweepstring
-        doTask.delay(task_desc)
-        response = {"response":"Job put in the task queue"}    
-        return jsonify(response)
-        
 @app.route('/GrapleRun', defaults={'filtername': None}, methods= ['GET', 'POST'])
 @app.route('/GrapleRun/<filtername>', methods= ['GET', 'POST'])
 def upload_file(filtername):
@@ -481,36 +472,12 @@ def upload_file(filtername):
         # should put the task in queue here and return.
         task_desc = "graple_run_batch"+"$"+dir_name+"$"+filename
         if (filtername): 
+            filtername += '.R'
             doTask.delay(task_desc, filtername)
         else:  
             doTask.delay(task_desc)  
         return jsonify(response)
 
-
-@app.route('/GrapleRunMetOffset', defaults={'filtername': None}, methods= ['GET', 'POST'])
-@app.route('/GrapleRunMetOffset/<filtername>', methods= ['GET', 'POST'])
-def run_sweep(filtername):
-    global base_upload_path
-    if request.method == 'POST':
-        f = request.files['files']
-        filename = f.filename
-        response = {"uid":batch_id_generator()}
-        dir_name = os.path.join(base_upload_path,response["uid"])
-        os.mkdir(dir_name)
-        base_folder = os.path.join(dir_name,'base_folder')
-        os.mkdir(base_folder)
-        topdir = dir_name
-        os.chdir(base_folder)
-        f.save(filename)
-        os.chdir(topdir)
-        copy_tree(base_graple_path,topdir)
-        subprocess.call(['python' , 'CreateWorkingFolders.py'])
-        if(filtername):      
-            os.chdir("Scripts")
-    	    shutil.copy(os.path.join(topdir,filtername),os.getcwd())      
-            os.chdir(topdir)
-        return jsonify(response)
-      
 @app.route('/GrapleRunResults/<uid>', methods=['GET','POST'])
 def return_consolidated_output(uid):
     global base_upload_path
@@ -523,6 +490,9 @@ def return_consolidated_output(uid):
         if not status["curr_status"].startswith("100"):
             ret_dict["status"]="Job under processing,please try agian after some time."
             return jsonify(ret_dict)
+        summary_file = os.path.join(dir_name,"base_folder","sim_summary.csv")
+        if os.path.isfile(summary_file):
+            shutil.copy(summary_file,os.path.join(dir_name,'Results'))
         process_graple_results(dir_name)
         #output_file = os.path.join(uid,'Results','output.zip')
         output_file = os.path.join(uid,'Results','output.tar.gz')
@@ -530,33 +500,6 @@ def return_consolidated_output(uid):
         ret_dict["output_url"]=url
         ret_dict["status"]="success"
         return jsonify(ret_dict)
-        
-@app.route('/GrapleRunResultsMetSample/<uid>', methods=['GET','POST'])
-def return_distributionJob_consolidated_output(uid):
-    global base_upload_path
-    ret_dict = {}
-    if request.method == 'GET':
-        dir_name = os.path.join(base_upload_path,uid)
-        # Sanity check to ensure that job processing is completed.
-        status = {}
-        status["curr_status"] = check_Job_status(dir_name)
-        if not status["curr_status"].startswith("100"):
-            ret_dict["status"]="Job under processing,please try agian after some time."
-            return jsonify(ret_dict)
-        #result_zip = os.path.join(dir_name,'Results','output.zip')
-        summary_file_path = os.path.join(dir_name,"base_folder","sim_summary.csv")
-        # club summary_file to the results
-        shutil.copy(summary_file_path,os.path.join(dir_name,'Results'))
-        process_graple_results(dir_name)
-        # use gunzip to unzip
-        # 
-        #subprocess.call(['zip' ,'-r',result_zip,'sim_summary.csv'])
-        output_file=os.path.join(uid,'Results','output.tar.gz')
-        url = url_for('static',filename=output_file)
-        ret_dict["output_url"]=url
-        ret_dict["status"]="success"
-        return jsonify(ret_dict)
-
         
 @app.route('/download_file/<request_string>', methods=['GET','POST'])
 def return_requested_file(request_string):
@@ -597,7 +540,7 @@ def get_PPOLibrary_scripts():
     if request.method == 'GET':
         scriptsDir = os.path.join(base_graple_path, "Filters")
         if(os.path.exists(scriptsDir)):
-            filesList = os.listdir(scriptsDir) 
+            filesList = [os.path.splitext(filtername)[0] for filtername in os.listdir(scriptsDir)]
     return json.dumps(filesList)        
 
 @app.route('/GrapleGetVersion', methods=['GET'])
